feat: update BatchCallablePreparedStatement
