+++ java/openmldb-jdbc/src/main/java/com/_4paradigm/openmldb/jdbc/PreparedStatement.java
@@ -17,51 +17,21 @@
 package com._4paradigm.openmldb.jdbc;
 
 import com._4paradigm.openmldb.*;
-import com._4paradigm.openmldb.sdk.Common;
-import com._4paradigm.openmldb.sdk.impl.Util;
 
 import java.io.InputStream;
 import java.io.Reader;
 import java.math.BigDecimal;
 import java.net.URL;
-import java.nio.charset.Charset;
-import java.nio.charset.StandardCharsets;
 import java.sql.Date;
 import java.sql.ResultSet;
 import java.sql.*;
 import java.util.*;
 
-public class PreparedStatement implements java.sql.PreparedStatement {
-    public static final Charset CHARSET = StandardCharsets.UTF_8;
-    protected String db;
-    protected String currentSql;
+public abstract class PreparedStatement implements java.sql.PreparedStatement {
     protected SQLRouter router;
-    protected SQLRequestRow currentRow;
-    protected Schema currentSchema;
-    protected TreeMap<Integer, com._4paradigm.openmldb.DataType> types;
-    protected TreeMap<Integer, com._4paradigm.openmldb.DataType> orgTypes;
-    protected TreeMap<Integer, Object> currentDatas;
     protected boolean closed = false;
     protected boolean closeOnComplete = false;
-    protected Map<Integer, Integer> stringsLen = new HashMap<>();
 
-    private void checkNull() throws SQLException {
-        if (db == null) {
-            throw new SQLException("db is null");
-        }
-        if (currentSql == null) {
-            throw new SQLException("sql is null");
-        }
-        if (router == null) {
-            throw new SQLException("SQLRouter is null");
-        }
-        if (currentDatas == null) {
-            throw new SQLException("currentDatas is null");
-        }
-        if (types == null) {
-            throw new SQLException("currentDatas is null");
-        }
-    }
 
     protected void checkClosed() throws SQLException {
         if (closed) {
@@ -75,62 +45,10 @@ public class PreparedStatement implements java.sql.PreparedStatement {
         }
     }
 
-    void checkIdx(int i) throws SQLException {
-        checkClosed();
-        checkNull();
-        if (i <= 0) {
-            throw new SQLException("index out of array");
-        }
-        if (currentDatas.containsKey(i)) {
-            throw new SQLException("index duplicate, index: " + i + " already exist");
-        }
-    }
-
-    @Override
-    public SQLResultSet executeQuery() throws SQLException {
-        checkClosed();
-        checkExecutorClosed();
-        dataBuild();
-        Status status = new Status();
-        com._4paradigm.openmldb.ResultSet resultSet = router.ExecuteSQLParameterized(db, currentSql, currentRow, status);
-        if (resultSet == null || status.getCode() != 0) {
-            String msg = status.ToString();
-            status.delete();
-            if (resultSet != null) {
-                resultSet.delete();
-            }
-            throw new SQLException("execute sql fail, msg: " + msg);
-        }
-        status.delete();
-        SQLResultSet rs = new SQLResultSet(resultSet);
-        if (closeOnComplete) {
-            closed = true;
-        }
-        return rs;
-    }
-
     @Override
     @Deprecated
     public int executeUpdate() throws SQLException {
         throw new SQLException("current do not support this method");
-    }
-
-    @Override
-    public void setNull(int parameterIndex, int sqlType) throws SQLException {
-        setNull(parameterIndex, Util.sqlTypeToDataType(sqlType));
-    }
-
-    private void setNull(int i, DataType type) throws SQLException {
-        checkIdx(i);
-        types.put(i, type);
-        currentDatas.put(i, null);
-    }
-
-    @Override
-    public void setBoolean(int i, boolean b) throws SQLException {
-        checkIdx(i);
-        types.put(i, DataType.kTypeBool);
-        currentDatas.put(i, b);
     }
 
     @Override
@@ -140,58 +58,11 @@ public class PreparedStatement implements java.sql.PreparedStatement {
     }
 
     @Override
-    public void setShort(int i, short i1) throws SQLException {
-        checkIdx(i);
-        types.put(i, DataType.kTypeInt16);
-        currentDatas.put(i, i1);
-    }
-
-    @Override
-    public void setInt(int i, int i1) throws SQLException {
-        checkIdx(i);
-        types.put(i, DataType.kTypeInt32);
-        currentDatas.put(i, i1);
-    }
-
-    @Override
-    public void setLong(int i, long l) throws SQLException {
-        checkIdx(i);
-        types.put(i, DataType.kTypeInt64);
-        currentDatas.put(i, l);
-    }
-
-    @Override
-    public void setFloat(int i, float v) throws SQLException {
-        checkIdx(i);
-        types.put(i, DataType.kTypeFloat);
-        currentDatas.put(i, v);
-    }
-
-    @Override
-    public void setDouble(int i, double v) throws SQLException {
-        checkIdx(i);
-        types.put(i, DataType.kTypeDouble);
-        currentDatas.put(i, v);
-    }
-
-    @Override
     @Deprecated
     public void setBigDecimal(int i, BigDecimal bigDecimal) throws SQLException {
         throw new SQLException("current do not support this type");
     }
 
-    @Override
-    public void setString(int i, String s) throws SQLException {
-        checkIdx(i);
-        if (s == null) {
-            setNull(i, DataType.kTypeString);
-            return;
-        }
-        types.put(i, DataType.kTypeString);
-        byte[] bytes = s.getBytes(CHARSET);
-        stringsLen.put(i, bytes.length);
-        currentDatas.put(i, bytes);
-    }
 
     @Override
     @Deprecated
@@ -200,32 +71,9 @@ public class PreparedStatement implements java.sql.PreparedStatement {
     }
 
     @Override
-    public void setDate(int i, Date date) throws SQLException {
-        checkIdx(i);
-        if (date == null) {
-            setNull(i, DataType.kTypeDate);
-            return;
-        }
-        types.put(i, DataType.kTypeDate);
-        currentDatas.put(i, date);
-    }
-
-    @Override
     @Deprecated
     public void setTime(int i, Time time) throws SQLException {
         throw new SQLException("current do not support this type");
-    }
-
-    @Override
-    public void setTimestamp(int i, Timestamp timestamp) throws SQLException {
-        checkIdx(i);
-        if (timestamp == null) {
-            setNull(i, DataType.kTypeTimestamp);
-            return;
-        }
-        types.put(i, DataType.kTypeTimestamp);
-        long ts = timestamp.getTime();
-        currentDatas.put(i, ts);
     }
 
     @Override
@@ -247,88 +95,11 @@ public class PreparedStatement implements java.sql.PreparedStatement {
     }
 
     @Override
-    public void clearParameters() {
-        currentDatas.clear();
-        types.clear();
-        stringsLen.clear();
-    }
-
-    @Override
     @Deprecated
     public void setObject(int i, Object o, int i1) throws SQLException {
         throw new SQLException("current do not support this method");
     }
 
-    protected void dataBuild() throws SQLException {
-        if (types == null) {
-            throw new SQLException("fail to build data when data types is null");
-        }
-        // types has been updated
-        if (null == this.currentRow || orgTypes != types) {
-            if (types.firstKey() != 1 || types.lastKey() != types.size()) {
-                throw new SQLException("data not enough, indexes are " + currentDatas.keySet());
-            }
-            ColumnTypes columnTypes = new ColumnTypes();
-            for (int i = 0; i < types.size(); i++) {
-                columnTypes.AddColumnType(types.get(i + 1));
-            }
-            this.currentRow = SQLRequestRow.CreateSQLRequestRowFromColumnTypes(columnTypes);
-            if (this.currentRow == null) {
-        	throw new SQLException("fail to create sql request row from column types");
-            }
-            this.currentSchema = this.currentRow.GetSchema();
-            this.orgTypes = this.types;
-        }
-        if (this.currentSchema == null) {
-            throw new SQLException("fail to build data with null schema");
-        }
-        int strLen = 0;
-        for (Map.Entry<Integer, Integer> entry : stringsLen.entrySet()) {
-            strLen += entry.getValue();
-        }
-        boolean ok = this.currentRow.Init(strLen);
-        if (!ok) {
-            throw new SQLException("build data row failed");
-        }
-        for (int i = 0; i < this.currentSchema.GetColumnCnt(); i++) {
-            DataType dataType = this.currentSchema.GetColumnType(i);
-            Object data = this.currentDatas.get(i + 1);
-            if (data == null) {
-                ok = this.currentRow.AppendNULL();
-            } else {
-                if (DataType.kTypeBool.equals(dataType)) {
-                    ok = this.currentRow.AppendBool((boolean) data);
-                } else if (DataType.kTypeDate.equals(dataType)) {
-                    Date date = (Date) data;
-                    ok = this.currentRow.AppendDate(date.getYear() + 1900, date.getMonth() + 1, date.getDate());
-                } else if (DataType.kTypeDouble.equals(dataType)) {
-                    ok = this.currentRow.AppendDouble((double) data);
-                } else if (DataType.kTypeFloat.equals(dataType)) {
-                    ok = this.currentRow.AppendFloat((float) data);
-                } else if (DataType.kTypeInt16.equals(dataType)) {
-                    ok = this.currentRow.AppendInt16((short) data);
-                } else if (DataType.kTypeInt32.equals(dataType)) {
-                    ok = this.currentRow.AppendInt32((int) data);
-                } else if (DataType.kTypeInt64.equals(dataType)) {
-                    ok = this.currentRow.AppendInt64((long) data);
-                } else if (DataType.kTypeString.equals(dataType)) {
-                    byte[] bdata = (byte[]) data;
-                    ok = this.currentRow.AppendString(bdata, bdata.length);
-                } else if (DataType.kTypeTimestamp.equals(dataType)) {
-                    ok = this.currentRow.AppendTimestamp((long) data);
-                } else {
-                    throw new SQLException("unkown data type " + dataType.toString());
-                }
-            }
-            if (!ok) {
-                throw new SQLException("append data failed, idx is " + i);
-            }
-        }
-        if (!this.currentRow.Build()) {
-            throw new SQLException("build request row failed");
-        }
-        clearParameters();
-    }
 
     @Override
     @Deprecated
@@ -376,13 +147,6 @@ public class PreparedStatement implements java.sql.PreparedStatement {
     @Deprecated
     public void setArray(int i, Array array) throws SQLException {
         throw new SQLException("current do not support this method");
-    }
-
-    @Override
-    public ResultSetMetaData getMetaData() throws SQLException {
-        checkClosed();
-        checkNull();
-        return new SQLResultSetMetaData(Common.convertSchema(this.currentSchema));
     }
 
     @Override
@@ -549,21 +313,7 @@ public class PreparedStatement implements java.sql.PreparedStatement {
 
     @Override
     public void close() throws SQLException {
-        this.db = null;
-        this.currentSql = null;
         this.router = null;
-        if (this.currentSchema != null) {
-            this.currentSchema.delete();
-            this.currentSchema = null;
-        }
-        this.currentDatas = null;
-        this.types = null;
-        this.orgTypes = null;
-        this.stringsLen = null;
-        if (this.currentRow != null) {
-            this.currentRow.delete();
-            this.currentRow = null;
-        }
         this.closed = true;
     }
