+++ java/openmldb-jdbc/src/main/java/com/_4paradigm/openmldb/sdk/impl/CallablePreparedStatementImpl.java
@@ -19,6 +19,7 @@ package com._4paradigm.openmldb.sdk.impl;
 import com._4paradigm.openmldb.SQLRouter;
 import com._4paradigm.openmldb.Status;
 import com._4paradigm.openmldb.common.codec.CodecUtil;
+import com._4paradigm.openmldb.common.codec.FlexibleRowBuilder;
 import com._4paradigm.openmldb.jdbc.CallablePreparedStatement;
 import com._4paradigm.openmldb.jdbc.SQLResultSet;
 import com._4paradigm.openmldb.jdbc.SQLResultSetMetaData;
@@ -34,8 +35,9 @@ public class CallablePreparedStatementImpl extends CallablePreparedStatement {
     private int routerCol;
     private String routerValue = "";
 
-    public CallablePreparedStatementImpl(String db, Deployment deployment, SQLRouter router) throws SQLException {
-        super(db, deployment, router);
+    public CallablePreparedStatementImpl(Deployment deployment, SQLRouter router) throws SQLException {
+        super(deployment, router);
+        rowBuilder = new FlexibleRowBuilder(deployment.getInputMetaData(), true);
         routerCol = deployment.getRouterCol();
     }
 
@@ -60,6 +62,7 @@ public class CallablePreparedStatementImpl extends CallablePreparedStatement {
         }
         status.delete();
         SQLResultSet rs = new SQLResultSet(resultSet, deployment.getOutputSchema());
+        clearParameters();
         if (closeOnComplete) {
             closed = true;
         }
@@ -86,12 +89,14 @@ public class CallablePreparedStatementImpl extends CallablePreparedStatement {
             throw new SQLException("call procedure fail, msg: " + msg);
         }
         status.delete();
+        clearParameters();
         return new QueryFuture(queryFuture);
     }
 
     @Override
-    public ResultSetMetaData getMetaData() {
-        return new SQLResultSetMetaData(deployment.getInputSchema());
+    public void clearParameters() {
+        rowBuilder.clear();
+        routerValue = "";
     }
 
     @Override
